修改评分类型

+++ chunyu-cms-web/server/db/schema/movie/rate.ts
@@ -1,10 +1,10 @@
-import { int, mysqlTable } from 'drizzle-orm/mysql-core';
+import { int, mysqlTable, float } from 'drizzle-orm/mysql-core';
 import { columnsHelpers } from '../../columns.helpers';
 
 export const movieRateTable = mysqlTable('movie_rate', {
   movieRateId: int('movie_rate_id').autoincrement().primaryKey(),
   movieBasicsId: int('movie_basics_id').notNull().unique(),
-  rate: int('rate').notNull(),
+  rate: float('rate').notNull(),
   rateUserCount: int('rate_user_count').notNull(),
   ...columnsHelpers
 });
